Update: Refactor of ApiFootball Connection Service Refactored ApiFootball Connection service to handle only connecting to API Added Services for Fixtures, Leagues, Seasons and Teams as part of API Footall Module.
